+++ docs/source/control/im.rst
@@ -2,7 +2,7 @@
 Induction Machine
 *****************
 
-Control methods for induction machines will be described here, but so far only reduced-order observers are documented...
+Control methods for induction machines will be described here. So far only reduced-order observers are documented. This section will be extended and improved in the future.
 
 .. toctree::
    :titlesonly:
