+++ docs/source/control/sm.rst
@@ -2,5 +2,5 @@
 Synchronous Machines
 ********************
 
-Control methods for synchronous machines will be described here, nothing yet...
+Control methods for synchronous machines will be described here in the future...
