+++ examples/obs_vhz/plot_obs_vhz_ctrl_im_2kw.py
@@ -8,13 +8,11 @@ drive.
 """
 
 # %%
-# Import the packages.
+# Imports.
 
 import numpy as np
-import motulator.model as model
-import motulator.control as control
-from motulator.helpers import BaseValues, Sequence
-from motulator.plots import plot
+from motulator import model, control
+from motulator import BaseValues, Sequence, plot
 
 # %%
 # Compute base values based on the nominal values (just for figures).
