+++ examples/obs_vhz/plot_obs_vhz_ctrl_pmsm_2kw.py
@@ -7,13 +7,11 @@ This example simulates observer-based V/Hz control of a 2.2-kW PMSM drive.
 """
 
 # %%
-# Import the package.
+# Imports.
 
 import numpy as np
-import motulator.model as model
-import motulator.control as control
-from motulator.helpers import BaseValues, Sequence
-from motulator.plots import plot
+from motulator import model, control
+from motulator import BaseValues, Sequence, plot
 
 # %%
 # Compute base values based on the nominal values (just for figures).
