+++ examples/obs_vhz/plot_obs_vhz_ctrl_pmsm_2kw_two_mass.py
@@ -11,14 +11,12 @@ example.
 """
 
 # %%
-# Import the packages.
+# Imports.
 
 import numpy as np
 import matplotlib.pyplot as plt
-import motulator.model as model
-import motulator.control as control
-from motulator.helpers import BaseValues, Sequence
-from motulator.plots import plot
+from motulator import model, control
+from motulator import BaseValues, Sequence, plot
 
 # %%
 # Compute base values based on the nominal values (just for figures).
