+++ examples/obs_vhz/plot_obs_vhz_ctrl_pmsyrm_thor.py
@@ -19,15 +19,13 @@ into account in the control algorithm.
 """
 
 # %%
-# Import the packages.
+# Imports.
 
 import numpy as np
 from scipy.optimize import minimize_scalar
 from scipy.interpolate import LinearNDInterpolator
-import motulator.model as model
-import motulator.control as control
-from motulator.helpers import BaseValues, Sequence
-from motulator.plots import plot
+from motulator import model, control
+from motulator import BaseValues, Sequence, plot
 
 # %%
 # Compute base values based on the nominal values (just for figures).
