+++ examples/signal_inj/plot_signal_inj_pmsm_2kw.py
@@ -8,14 +8,12 @@ Square-wave signal injection is used with a simple phase-locked loop.
 """
 
 # %%
-# Import the packages.
+# Imports.
 
 import numpy as np
 import matplotlib.pyplot as plt
-import motulator.model as model
-import motulator.control as control
-from motulator.helpers import BaseValues, Sequence
-from motulator.plots import plot
+from motulator import model, control
+from motulator import BaseValues, Sequence, plot
 
 # %%
 # Compute base values based on the nominal values (just for figures).
