+++ examples/vector/README.txt
@@ -1,2 +1,4 @@
 Vector Control
 --------------
+
+These examples are for current vector control of induction machines and synchronous machines. The magnetic saturation model of an induction machine is also demonstrated (:doc:`/auto_examples/vector/plot_vector_ctrl_im_2kw`) as well as computation of control look-up tables for synchronous machines (:doc:`/auto_examples/vector/plot_vector_ctrl_pmsyrm_thor`).
