+++ examples/vector/plot_vector_ctrl_pmsyrm_thor.py
@@ -8,13 +8,12 @@ synchronous reluctance motor. Control look-up tables are also plotted.
 """
 
 # %%
-# Import the packages.
+# %%
+# Imports.
 
 import numpy as np
-import motulator.model as model
-import motulator.control as control
-from motulator.helpers import BaseValues
-from motulator.plots import plot
+from motulator import model, control
+from motulator import BaseValues, plot
 
 # %%
 # Compute base values based on the nominal values (just for figures).
