+++ examples/vector/plot_vector_ctrl_syrm_7kw.py
@@ -7,13 +7,12 @@ This example simulates sensorless vector control of a 6.7-kW SyRM drive.
 """
 
 # %%
-# Import the packages.
+# %%
+# Imports.
 
 import numpy as np
-import motulator.model as model
-import motulator.control as control
-from motulator.helpers import BaseValues, Sequence
-from motulator.plots import plot
+from motulator import model, control
+from motulator import BaseValues, Sequence, plot
 
 # %%
 # Compute base values based on the nominal values (just for figures).
