+++ examples/vhz/README.txt
@@ -1,2 +1,4 @@
 V/Hz Control
 ------------
+
+These examples shows operation of an induction machine under open-loop V/Hz control. Furthermore, a diode front-end rectifier and transition to six-step modulation are also demonstrated.
