+++ motulator/__init__.py
@@ -1,0 +1,28 @@
+"""
+*motulator*: Motor drive simulator in Python
+
+This software includes continuous-time simulation models for induction machines 
+and synchronous machines. Furthermore, selected examples of discrete-time 
+control algortihms are also included as well as various utilities.
+
+"""
+
+from motulator._helpers import (
+    abc2complex, complex2abc, BaseValues, Sequence, Step)
+
+from motulator._plots import plot, plot_extra
+
+from motulator import control
+from motulator import model
+
+__all__ = [
+    "abc2complex",
+    "complex2abc",
+    "BaseValues",
+    "Sequence",
+    "Step",
+    "control",
+    "model",
+    "plot",
+    "plot_extra",
+]
