+++ motulator/_plots.py
@@ -4,7 +4,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
 from cycler import cycler
-from motulator.helpers import complex2abc
+from motulator._helpers import complex2abc
 from motulator._utils import Bunch
 
 # Plotting parameters
