+++ motulator/control/im/_vhz.py
@@ -17,7 +17,7 @@ References
 # %%
 import numpy as np
 from motulator.control._common import Ctrl, PWM
-from motulator.helpers import abc2complex
+from motulator._helpers import abc2complex
 from motulator._utils import Bunch
