+++ motulator/model/im/_drive.py
@@ -6,7 +6,7 @@ implemented in stator coordinates.
 
 """
 import numpy as np
-from motulator.helpers import abc2complex, complex2abc
+from motulator._helpers import abc2complex, complex2abc
 from motulator._utils import Bunch
